# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pixelate",
        "description": "Documentation for Pixelate app",
        "version": "1.0.2"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Server"
        }
    ],
    "paths": {
...
        "/api/designs": {
            "get": {
                "tags": [
                    "Design"
                ],
                "summary": "Get designs that are Live",
                "description": "Get all designs available online (set to live )",
                "operationId": "App\\Http\\Controllers\\Designs\\DesignController::index",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Design"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "UDesign Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Design"
                ],
                "summary": "Upload Image",
                "description": "Uploads Image Local or to S3 bucket",
                "operationId": "App\\Http\\Controllers\\Designs\\UploadController::upload",
                "requestBody": {
                    "description": "Upload images request body",
                    "required": true,
                    "content": {
                        "application/octet-stream": {
                            "schema": {
                                "type": "string",
                                "format": "binary"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Design"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "image": {
                                                    "example": [
                                                        "The image field is required."
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
        "/api/image/{id}": {
            "get": {
                "tags": [
                    "Image"
                ],
                "summary": "Get Image",
                "description": "Get image after upload",
                "operationId": "App\\Http\\Controllers\\Designs\\UploadController::getImage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "image": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "UDesign Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/users": {
            "get": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Retrieve All Users information",
                "description": "Gets alkl users from DB  ** Reqiures Authorisation: Add Auth heather by clicking the Lock icon above",
                "operationId": "App\\Http\\Controllers\\User\\UserController::index",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/UserProfile"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User should be authorized to get profile information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not authorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Retrieve All Users information",
                "description": "Gets alkl users from DB  ** Reqiures Authorisation: Add Auth heather by clicking the Lock icon above",
                "operationId": "App\\Http\\Controllers\\User\\UserController::findUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/UserProfile"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User should be authorized to get profile information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not authorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/user/{username}": {
            "get": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Retrieve All Users information",
                "description": "Gets alkl users from DB  ** Reqiures Authorisation: Add Auth heather by clicking the Lock icon above",
                "operationId": "App\\Http\\Controllers\\User\\UserController::findByUserName",
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "pierce"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/UserProfile"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User should be authorized to get profile information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not authorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "UserProfile": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/User"
                    }
                ]
            },
...
            "Design": {
                "properties": {
                    "id": {
                        "description": "App\\Models\\Design",
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "title": {
                        "description": "App\\Models\\Design",
                        "type": "string",
                        "example": "the sea life"
                    },
                    "slug": {
                        "description": "App\\Models\\Design",
                        "type": "string",
                        "example": "the-sea-life"
                    },
                    "disk": {
                        "description": "App\\Models\\Design",
                        "type": "string",
                        "example": "public"
                    },
                    "is_live": {
                        "description": "App\\Models\\Design",
                        "type": "boolean",
                        "example": 0
                    },
                    "images": {
                        "description": "App\\Models\\Design",
                        "properties": {
                            "thumbnail": {
                                "description": "App\\Models\\Design",
                                "type": "string",
                                "example": "http://localhost:8000/storage/uploads/designs/thumbnail/"
                            },
                            "large": {
                                "description": "App\\Models\\Design",
                                "type": "string",
                                "example": "http://localhost:8000/storage/uploads/designs/large/"
                            },
                            "original": {
                                "description": "App\\Models\\Design",
                                "type": "string",
                                "example": "http://localhost:8000/storage/uploads/designs/original/"
                            }
                        },
                        "type": "object"
                    },
                    "likes_count": {
                        "description": "App\\Models\\Design",
                        "type": "integer",
                        "example": 0
                    },
                    "uploaded_successful": {
                        "description": "App\\Models\\Design",
                        "type": "interger",
                        "example": null
                    },
                    "tag_list": {
                        "description": "App\\Models\\Design",
                        "properties": {
                            "tag": {
                                "description": "List of Tags added to design, array of many tags",
                                "example": [
                                    "Cool",
                                    "Bright Day"
                                ]
                            },
                            "normalized": {
                                "description": "List of Tags normalized, lowercase and slugged",
                                "example": [
                                    "cool",
                                    "bright-day"
                                ]
                            }
                        },
                        "type": "object"
                    },
                    "created_dates": {
                        "description": "App\\Models\\Design",
                        "properties": {
                            "created_at_human": {
                                "description": "Date Created formatted",
                                "example": "52 minutes ago"
                            },
                            "created_at": {
                                "description": "Raw unfarmatted Date ",
                                "example": "2020-11-09T20:04:11.000000Z"
                            }
                        },
                        "type": "object"
                    },
                    "updated_dates": {
                        "description": "App\\Models\\Design",
                        "properties": {
                            "updated_at_human": {
                                "description": "Date Updated formatted",
                                "example": "52 minutes ago"
                            },
                            "updated_at": {
                                "description": "Raw unfarmatted update date ",
                                "example": "2020-11-09T20:04:11.000000Z"
                            }
                        },
                        "type": "object"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "comment": {
                        "$ref": "#/components/schemas/Comment"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Design"
                }
            },
...
            "User": {
                "required": [
                    "password"
                ],
                "properties": {
                    "id": {
                        "description": "App\\Models\\User",
                        "type": "integer",
                        "readOnly": "true",
                        "example": 1
                    },
                    "name": {
                        "description": "App\\Models\\User",
                        "type": "string",
                        "maxLength": 32,
                        "example": "Mittie Morissette"
                    },
                    "username": {
                        "description": "App\\Models\\User",
                        "type": "string",
                        "maxLength": 32,
                        "example": "pierce"
                    },
                    "email": {
                        "description": "User unique email address",
                        "type": "string",
                        "format": "email",
                        "readOnly": "true",
                        "example": "fanny256@email.com"
                    },
                    "email_verified_at": {
                        "description": "Datetime marker of verification status",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true",
                        "example": "2019-02-25 12:59:20"
                    },
                    "two_factor_secret": {
                        "description": "App\\Models\\User",
                        "type": "number",
                        "example": null
                    },
                    "two_factor_recovery_codes": {
                        "description": "App\\Models\\User",
                        "type": "number",
                        "example": null
                    },
                    "tagline:": {
                        "description": "App\\Models\\User",
                        "type": "string",
                        "maxLength": 32,
                        "example": "Producer:"
                    },
                    "location": {
                        "description": "App\\Models\\User",
                        "properties": {
                            "type": {
                                "description": "App\\Models\\User",
                                "type": "string",
                                "example": "point"
                            },
                            "coordinates": {
                                "description": "App\\Models\\User",
                                "example": [
                                    8.503972,
                                    51.017243
                                ]
                            }
                        },
                        "type": "object"
                    },
                    "formatted_address": {
                        "description": "App\\Models\\User",
                        "example": "811 Sibyl Bypass Suite 783\\n New Rita, AL 48220-0930"
                    },
                    "available_to_hire": {
                        "description": "App\\Models\\User",
                        "type": "boolean",
                        "example": 1
                    },
                    "about": {
                        "description": "App\\Models\\User",
                        "type": "string",
                        "example": "VERY deeply with a soldier on each."
                    },
                    "trade_id": {
                        "description": "App\\Models\\User",
                        "type": "number",
                        "example": null
                    },
                    "current_team_id": {
                        "description": "App\\Models\\User",
                        "type": "number",
                        "example": null
                    },
                    "profile_photo_path": {
                        "description": "App\\Models\\User",
                        "type": "number",
                        "example": null
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "updated_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    },
                    "photo_url:": {
                        "description": "App\\Models\\User",
                        "type": "string",
                        "maxLength": 32,
                        "example": "https://www.gravatar.com/avatar/97bd1823e00f02eb71a0b709425152a7jpg?s=200&d=mm:"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            }
        },
        "securitySchemes": {
            "token": {
                "type": "apiKey",
                "description": "Example of Auth Header Token (Copy and paste below to authorize)=>\nBearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC9sb2NhbGhvc3Q6ODAwMFwvYXBpXC9sb2dpbiIsImlhdCI6MTYwNTM0ODU1OSwiZXhwIjoxNjA1MzUyMTU5LCJuYmYiOjE2MDUzNDg1NTksImp0aSI6IjlqMDN0MVVISThoNjlhRUsiLCJzdWIiOjEsInBydiI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzllNzAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.dn45EdZMatpW2zAQcZklDS2902mwVNx44TbkDiiGLXY",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}